+++ examples/resources/sdwan_transport_management_vpn_profile_parcel/resource.tf
@@ -18,7 +18,7 @@ resource "sdwan_transport_management_vpn_profile_parcel" "example" {
       network_address = "1.2.3.4"
       subnet_mask     = "0.0.0.0"
       gateway         = "nextHop"
-      ipv4_route_gateway_next_ho = [
+      next_hops = [
         {
           address                 = "1.2.3.4"
           administrative_distance = 1