+++ internal/provider/data_source_sdwan_transport_management_vpn_profile_parcel_test.go
@@ -42,8 +42,8 @@ func TestAccDataSourceSdwanTransportManagementVPNProfileParcel(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv4_static_routes.0.network_address", "1.2.3.4"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv4_static_routes.0.subnet_mask", "0.0.0.0"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv4_static_routes.0.gateway", "nextHop"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv4_static_routes.0.ipv4_route_gateway_next_ho.0.address", "1.2.3.4"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv4_static_routes.0.ipv4_route_gateway_next_ho.0.administrative_distance", "1"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv4_static_routes.0.next_hops.0.address", "1.2.3.4"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv4_static_routes.0.next_hops.0.administrative_distance", "1"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv4_static_routes.0.administrative_distance", "1"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv6_static_routes.0.prefix", "2002::/16"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.sdwan_transport_management_vpn_profile_parcel.test", "ipv6_static_routes.0.next_hops.0.address", "2001:0:0:1::1"))
@@ -92,7 +92,7 @@ func testAccDataSourceSdwanTransportManagementVPNProfileParcelConfig() string {
 	config += `	  network_address = "1.2.3.4"` + "\n"
 	config += `	  subnet_mask = "0.0.0.0"` + "\n"
 	config += `	  gateway = "nextHop"` + "\n"
-	config += `	  ipv4_route_gateway_next_ho = [{` + "\n"
+	config += `	  next_hops = [{` + "\n"
 	config += `		address = "1.2.3.4"` + "\n"
 	config += `		administrative_distance = 1` + "\n"
 	config += `	}]` + "\n"