+++ internal/provider/resource_sdwan_transport_management_vpn_profile_parcel.go
@@ -191,7 +191,7 @@ func (r *TransportManagementVPNProfileParcelResource) Schema(ctx context.Context
 								stringvalidator.OneOf("nextHop", "dhcp", "null0"),
 							},
 						},
-						"ipv4_route_gateway_next_ho": schema.ListNestedAttribute{
+						"next_hops": schema.ListNestedAttribute{
 							MarkdownDescription: helpers.NewAttributeDescription("IPv4 Route Gateway Next Hop").String,
 							Optional:            true,
 							NestedObject: schema.NestedAttributeObject{